--- src/components/forms/Form.css
@@ -1,137 +1,0 @@
-/* Form Layout */
-form {
-  max-width: 600px;
-  margin: 0 auto;
-  padding: 20px;
-  border: 1px solid #ddd;
-  border-radius: 8px;
-  background-color: #ffffff; /* White background to match post hover state */
-}
-
-form div {
-  margin-bottom: 15px;
-}
-
-label {
-  display: block;
-  font-weight: bold;
-  margin-bottom: 5px;
-}
-
-/* Ensure global styles are applied properly */
-input[type="text"],
-textarea {
-  width: 100%;
-  padding: 8px;
-  border: 1px solid #ccc;
-  border-radius: 4px;
-  font-size: 0.8rem;
-  box-sizing: border-box;
-}
-
-textarea {
-  height: 100px;
-  resize: vertical;
-}
-
-/* Button Colors */
-button {
-  padding: 10px 20px;
-  border-radius: 4px;
-  cursor: pointer;
-  font-size: 16px;
-  border: none; /* Remove border for consistent styling */
-  transition: background-color 0.3s, transform 0.3s; /* Smooth transition effects */
-}
-
-.new-brand-button,
-.add-brand-button {
-  background-color: #ffc107; /* Yellow background for add brand actions */
-  color: white;
-}
-
-.new-brand-button:hover,
-.add-brand-button:hover {
-  background-color: #e0a800; /* Darker yellow on hover */
-}
-
-.save-button {
-  background-color: #007bff; /* Blue background for save */
-  color: white;
-}
-
-.save-button:hover {
-  background-color: #0056b3; /* Darker blue on hover */
-}
-
-.cancel-button {
-  background-color: #f44336; /* Red background for cancel */
-  color: white;
-  margin-left: 10px;
-}
-
-.cancel-button:hover {
-  background-color: #d32f2f; /* Darker red on hover */
-}
-
-/* Modal Styles */
-.ReactModal__Overlay {
-  background-color: rgba(0, 0, 0, 0.5); /* Semi-transparent overlay */
-  z-index: 1000;
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-}
-
-.ReactModal__Content {
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  background-color: #f0f8ff;
-  border-radius: 4px;
-  padding: 20px;
-  max-width: 500px;
-  width: 100%;
-  box-sizing: border-box; /* Ensure padding is included in the width */
-}
-
-/* Ensure modal inputs are constrained */
-.ReactModal__Content input[type="text"],
-.ReactModal__Content textarea {
-  width: 100%; /* Full width inside modal */
-  box-sizing: border-box; /* Include padding and border in width */
-}
-
-.ReactModal__Content .modal-input {
-  width: 100%;
-  box-sizing: border-box; /* Include padding and border in width */
-}
-
-/* Brand input brand suggestions dropdown styling */
-.suggestions-list {
-  position: absolute;
-  background: white;
-  border: 1px solid #ccc;
-  max-height: 150px; /* Adjust as needed */
-  overflow-y: auto;
-  width: calc(100% - 20px); /* To match the width of the input field */
-  margin-top: 5px;
-  z-index: 1000; /* Ensure it appears above other elements */
-  list-style-type: none;
-  padding: 0;
-  margin: 0;
-  border-radius: 4px;
-  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
-}
-
-.suggestions-list li {
-  padding: 10px;
-  cursor: pointer;
-}
-
-.suggestions-list li:hover {
-  background: #f0f0f0;
-}
