+++ src/components/modals/EditWishlistModal.jsx
@@ -1,0 +1,110 @@
+import { useEffect, useState } from "react";
+import { WishlistModal } from "./WishlistModal.jsx";
+import { editWishlistPost } from "../services/wishlistService.js";
+
+export const EditWishlistModal = ({
+  isOpen,
+  onClose,
+  userId,
+  wishlistItem,
+  onUpdate,
+}) => {
+  const [year, setYear] = useState("");
+  const [notes, setNotes] = useState("");
+  const [error, setError] = useState("");
+
+  useEffect(() => {
+    if (wishlistItem) {
+      setYear(wishlistItem.year || "");
+      setNotes(wishlistItem.notes || "");
+    }
+  }, [wishlistItem]);
+
+  const handleEditWishlistItem = async () => {
+    const updatedItem = {
+      id: wishlistItem.id,
+      userId,
+      year: parseInt(year),
+      brandId: wishlistItem.brand.id,
+      model: wishlistItem.model,
+      notes,
+    };
+
+    try {
+      await editWishlistPost(updatedItem);
+      onUpdate();
+      onClose();
+    } catch (err) {
+      setError("Failed to update wishlist item.");
+    }
+  };
+
+  return (
+    <WishlistModal
+      isOpen={isOpen}
+      onClose={onClose}
+      footerButtons={
+        <button className="btn btn-primary" onClick={handleEditWishlistItem}>
+          Save Changes
+        </button>
+      }
+    >
+      <div className="modal-header">
+        <h5 className="modal-title">
+          <strong>Edit Wishlist Item</strong>
+        </h5>
+        <button
+          type="button"
+          className="btn-close"
+          aria-label="Close"
+          onClick={onClose}
+        ></button>
+      </div>
+      <div className="modal-body">
+        <div className="mb-3">
+          <label htmlFor="year" className="form-label">
+            <strong>Year:</strong>
+          </label>
+          <input
+            className="form-control"
+            type="text"
+            id="year"
+            value={year}
+            onChange={(event) => {
+              setYear(event.target.value);
+            }}
+            autoComplete="off"
+          />
+        </div>
+        <div className="mb-3">
+          <label className="form-label">
+            <strong>Brand:</strong>
+          </label>
+          <p>{wishlistItem.brand.name}</p>
+        </div>
+        <div className="mb-3">
+          <label className="form-label">
+            <strong>Model:</strong>
+          </label>
+          <p>{wishlistItem.model}</p>
+        </div>
+        <div className="mb-3">
+          <label htmlFor="notes" className="form-label">
+            <strong>Notes:</strong>
+          </label>
+          <input
+            className="form-control"
+            type="text"
+            id="notes"
+            value={notes}
+            onChange={(event) => {
+              setNotes(event.target.value);
+            }}
+            autoComplete="off"
+          />
+        </div>
+        {error && <p className="text-danger">{error}</p>}
+      </div>
+    </WishlistModal>
+  );
+};
