+++ src/components/posts/AllPosts.jsx
@@ -56,7 +56,7 @@ export const AllPosts = ({ currentUser }) => {
   }, [selectedCategory, selectedBrand, forSale, allPosts]);
 
   return (
-    <div className="container all-posts-container">
+    <div className="container all-posts-container my-4">
       <h1 className="text-center mb-4">All Posts</h1>
       <div className="row mb-4">
         <div className="col-md-12 col-lg-6 mb-3 mb-lg-0">
